// apps/webapp/spa/src/lib/api/generated/models/BrandAudience.ts
export interface BrandAudience {
    /**
     * 
     * @type {string}
     * @memberof BrandAudience
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof BrandAudience
     */
    name?: string;
    /**
     * 
     * @type {BrandAudienceAgeRange}
     * @memberof BrandAudience
     */
    ageRange?: BrandAudienceAgeRange;
    /**
     * 
     * @type {BrandAudienceGender}
     * @memberof BrandAudience
     */
    gender?: BrandAudienceGender;
    /**
     * 
     * @type {BrandAudienceIncomeRange}
     * @memberof BrandAudience
     */
    incomeRange?: BrandAudienceIncomeRange;
    /**
     * 
     * @type {Array<string>}
     * @memberof BrandAudience
     */
    painPoints?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof BrandAudience
     */
    objections?: Array<string>;
    /**
     * 
     * @type {Array<ContentChannelName>}
     * @memberof BrandAudience
     */
    channels?: Array<ContentChannelName>;
}