// apps/webapp/spa/src/lib/api/generated/models/BrandColor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BrandColor
 */
export interface BrandColor {
    /**
     * 
     * @type {string}
     * @memberof BrandColor
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof BrandColor
     */
    hexValue: string;
}

/**
 * Check if a given object implements the BrandColor interface.
 */
export function instanceOfBrandColor(value: object): value is BrandColor {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('hexValue' in value) || value['hexValue'] === undefined) return false;
    return true;
}

export function BrandColorFromJSON(json: any): BrandColor {
    return BrandColorFromJSONTyped(json, false);
}

export function BrandColorFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandColor {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'hexValue': json['hex_value'],
    };
}

export function BrandColorToJSON(json: any): BrandColor {
    return BrandColorToJSONTyped(json, false);
}

export function BrandColorToJSONTyped(value?: BrandColor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'hex_value': value['hexValue'],
    };
}