// apps/webapp/spa/src/lib/api/generated/models/PositioningBrandData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voz AI API
 * Voz AI marketing automation platform API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PositioningBrandData
 */
export interface PositioningBrandData {
    /**
     * 
     * @type {string}
     * @memberof PositioningBrandData
     */
    description?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof PositioningBrandData
     */
    pointsOfDifference?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof PositioningBrandData
     */
    pointsOfParity?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof PositioningBrandData
     */
    productDescription?: string;
}

/**
 * Check if a given object implements the PositioningBrandData interface.
 */
export function instanceOfPositioningBrandData(value: object): value is PositioningBrandData {
    return true;
}

export function PositioningBrandDataFromJSON(json: any): PositioningBrandData {
    return PositioningBrandDataFromJSONTyped(json, false);
}

export function PositioningBrandDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): PositioningBrandData {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'pointsOfDifference': json['points_of_difference'] == null ? undefined : json['points_of_difference'],
        'pointsOfParity': json['points_of_parity'] == null ? undefined : json['points_of_parity'],
        'productDescription': json['product_description'] == null ? undefined : json['product_description'],
    };
}

export function PositioningBrandDataToJSON(json: any): PositioningBrandData {
    return PositioningBrandDataToJSONTyped(json, false);
}

export function PositioningBrandDataToJSONTyped(value?: PositioningBrandData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'description': value['description'],
        'points_of_difference': value['pointsOfDifference'],
        'points_of_parity': value['pointsOfParity'],
        'product_description': value['productDescription'],
    };
}